MPSGE.extract_scalars

MPSGE.production_sectors

MPSGE.sectors

MPSGE.sectors

MPSGE.solve!

MPSGE.extract_scalars — Method

extract_scalars

Takes a variable and extracts it the sub-variables.

MPSGE.production_sectors — Method

production_sectors(m::MPSGEModel)

Return all sectors that have a corresponding production block. These are coming from a dictionary, so order is not guaranteed.

This is primarily used when generating constraints.

MPSGE.sectors — Method

sectors(C::Commodity)

Return only the sectors that have the input commodity in their production block.

This is an optimization in building the model as the structure is very sparse iterating over all sectors is expensive.

MPSGE.sectors — Method

sectors(m::MPSGEModel)

Return all sectors in a model

MPSGE.solve! — Method

solve!(m::abstract_mpsge_model; keywords)
Function to solve the model. Triggers the build if the model hasn't been built yet.

Example

julia> solve!(m, cumulative_iteration_limit=0)