MPSGE.cost_function

MPSGE.extract_scalars

MPSGE.generate_report

MPSGE.production_sectors

MPSGE.revenue_function

MPSGE.sectors

MPSGE.sectors

MPSGE.solve!

MPSGE.@commodities

MPSGE.@commodity

MPSGE.@consumer

MPSGE.@consumers

MPSGE.@demand

MPSGE.@endowment

MPSGE.@final_demand

MPSGE.@input

MPSGE.@output

MPSGE.@parameter

MPSGE.@parameters

MPSGE.@production

MPSGE.@sector

MPSGE.@sectors

MPSGE.cost_function — Method

cost_function(S::ScalarSector; virtual = false)
cost_function(S::ScalarSector, nest::Symbol; virtual = false)

Return a vector of cost functions for the given sector and nest. If nest is not provided return the cost function for input tree.

nest is the symbol representing the nest. This can also be the name of a commodity.

If virtual is true, return the virtual cost functions.

MPSGE.extract_scalars — Method

extract_scalars

Takes a variable and extracts it the sub-variables.

MPSGE.generate_report — Method

generate_report(M::MPSGEModel)

Returns a dataframe with three columns, the variable, the value and the margin. The product of the value and the marge should be zero, if not the model has a specification error.

This function is useful for debugging models that fail a calibrated benchmark. If the model fails the benchmark, look for non-zero margins in this report, as they will reveal the error.

MPSGE.production_sectors — Method

production_sectors(m::MPSGEModel)

Return all sectors that have a corresponding production block. These are coming from a dictionary, so order is not guaranteed.

This is primarily used when generating constraints.

MPSGE.revenue_function — Method

revenue_function(S::ScalarSector; virtual = false)    
revenue_function(S::ScalarSector, nest::Symbol; virtual = false)

Return a vector of revenue functions for the given sector and nest. If nest is not provided return the revenue function for input tree.

nest is the symbol representing the nest. This can also be the name of a commodity.

If virtual is true, return the virtual revenue functions.

MPSGE.sectors — Method

sectors(C::Commodity)

Return only the sectors that have the input commodity in their production block.

This is an optimization in building the model as the structure is very sparse iterating over all sectors is expensive.

MPSGE.sectors — Method

sectors(m::MPSGEModel)

Return all sectors in a model

MPSGE.solve! — Method

solve!(m::abstract_mpsge_model; keywords)
Function to solve the model. Triggers the build if the model hasn't been built yet.

Example

julia> solve!(m, cumulative_iteration_limit=0)

MPSGE.@commodities — Macro

@commodities(model, args...)

Adds multiple commodities to a model at once.

The model must be the first argument, and multiple variables can be added on multiple lines by wrapping them in a begin ... end block.

Example

julia> M = MPSGEModel();

julia> I = [:a,:b];

julia> @commodities(M, begin
            X[I]
            Y[I,I], (description = "This is an indexed variable")
            Z, (description = "Scalar variable")
        end);

Note

Keywords must be contained in parentheses as in the example above.

MPSGE.@commodity — Macro

@commodity(model, expr, kwargs...)

Create a commodity in the model described by the expr with keyword arguments kwargs.

commodity names

expr must be one of the forms:

A single symbol like @commodity(model, X)

A container like expression @commodity(model, X[I])

Keyword Arguments

index: Explicity sets the index. For example, this @commodity(model, X, index = [I]) is equivalent to @commodity(model, X[I]).

description: Set a description on a variable.

MPSGE.@consumer — Macro

@consumer(model, expr, kwargs...)

Create a consumer in the model described by the expr with keyword arguments kwargs.

consumer names

expr must be one of the forms:

A single symbol like @consumer(model, X)

A container like expression @consumer(model, X[I])

Keyword Arguments

index: Explicity sets the index. For example, this @consumer(model, X, index = [I]) is equivalent to @consumer(model, X[I]).

description: Set a description on a variable.

MPSGE.@consumers — Macro

@consumers(model, args...)

Adds multiple consumers to a model at once.

The model must be the first argument, and multiple variables can be added on multiple lines by wrapping them in a begin ... end block.

Example

julia> M = MPSGEModel();

julia> I = [:a,:b];

julia> @consumers(M, begin
            X[I]
            Y[I,I], (description = "This is an indexed variable")
            Z, (description = "Scalar variable")
        end);

Note

Keywords must be contained in parentheses as in the example above.

MPSGE.@demand — Macro

@demand(model, consumer, demand_flow_block, kwargs...)

Example

@demand(M, CONS, begin
    @final_demand(X, 10)
    @endowment(Y, 5)
    @endowment(Z, 10)
end)

MPSGE.@endowment — Macro

@endowment(commodity, quantity, kwargs...)

MPSGE.@final_demand — Macro

@final_demand(commodity, quantity, kwargs...)

MPSGE.@input — Macro

@input(commodity, quantity, nest, kwargs...)

This macro is soley used within a @production block to define the inputs for a sector. The commodity, quantity and nest arguments are required, where nest is an object created in the @production macro.

Keyword Arguments

- `taxes` associate a tax to the commodity. The syntax `taxes = [Tax(RA, .1), Tax(GOVT, .3)]` defines two taxes, one associated the consumer `RA` and another with `GOVT`. In this example the values are fixed, but they can also be parameters.
- `reference_price=1` Set the reference price of a commodity.

MPSGE.@output — Macro

@output(commodity, quantity, nest, kwargs...)

This macro is soley used within a @production block to define the outputs for a sector. The commodity, quantity and nest arguments are required, where nest is an object created in the @production macro.

Keyword Arguments

- `taxes` associate a tax to the commodity. The syntax `taxes = [Tax(RA, .1), Tax(GOVT, .3)]` defines two taxes, one associated the consumer `RA` and another with `GOVT`. In this example the values are fixed, but they can also be parameters.
- `reference_price=1` Set the reference price of a commodity.

MPSGE.@parameter — Macro

@parameter(model, expr, value, kwargs...)

Create a parameter in the model described by the expr with initial value and keyword arguments kwargs.

parameter names

expr must be one of the forms:

A single symbol like @parameter(model, X)

A container like expression @parameter(model, X[I])

values

The value can either be a single float or an array. Currently using an array is finicky and it's recommended to set the value with set_parameter_value after creation if there are difficulties.

Keyword Arguments

index: Explicity sets the index. For example, this @parameter(model, X, index = [I]) is equivalent to @parameter(model, X[I]).

description: Set a description on a variable.

MPSGE.@parameters — Macro

@parameters(model, block)

Pluralized version of @parameter.

MPSGE.@production — Macro

@production(model, sector, nestings, netputs)

Define a production for the sector in the model with given nestings and netputs.

sector

This is any ScalarSector in the model.

nestings

This is where the nesting structure is defined and the associated elasticities. At minimum you must declare at least two nests and elasticities, one for the elasticity of substitution (input) and one for the elasticity of transformation (output), by convention these are denoted s and t respectively, although any identifier may be used.

As a minimal example, [s=1, t=0] will set the s nest to have an elasticity of 1 and the t nest 0. Suppose you want a nest below s called va with an elasticity of 2, this is created with [s=1, t=0, va=>s=2]. The va points at its parent nest s and the elasticity follows. Nestings can be aribrarily deep, for example

[s=1, t=0, va=>s=2, dm=>s=1, d=>dm=2]

will have two nests below s and one below dm.

netputs

A netput is either an @input or an @output. The netputs get wrapped in a begin ... end block and each netput must be on its own line.

Examples

In the below example we define the production blocks for two sectors X and Y. This is a non-function example solely created to show syntax. The X sector only has the two require elasticities where as Y has a more interesting nesting structure. A tax is included in the Y production block.

julia> M = MPSGEModel();

julia> @sectors(M, begin
            X
            Y
        end);

julia> @commodities(M, begin
            PX
            PY
            PL
            PK
        end);

julia> @consumer(M, RA);

julia> @production(M, X, [s=1,t=0], begin
            @output(PX, 10, t)
            @input(PL, 5, s)
            @input(PK, 5, s)
        end);

julia> @production(M, Y, [s=2, t=1, va=>s=1], begin
            @output(PY, 15, t)
            @input(PX, 3, s)
            @input(PL, 4, va, taxes = [Tax(RA, .5)])
            @input(PK, 6, va)
        end);

For examples using indexed sectors and commodities we recommend looking at the WiNDC national model. This will be linked when the appropriate write-up is ready.

MPSGE.@sector — Macro

@sector(model, expr, kwargs...)

Create a sector in the model described by the expr with keyword arguments kwargs.

sector names

expr must be one of the forms:

A single symbol like @sector(model, X)

A container like expression @sector(model, X[I])

Keyword Arguments

index: Explicity sets the index. For example, this @sector(model, X, index = [I]) is equivalent to @sector(model, X[I]).

description: Set a description on a variable.

MPSGE.@sectors — Macro

@sectors(model, args...)

Adds multiple sectors to a model at once.

The model must be the first argument, and multiple variables can be added on multiple lines by wrapping them in a begin ... end block.

Example

julia> M = MPSGEModel();

julia> I = [:a,:b];

julia> @sectors(M, begin
            X[I]
            Y[I,I], (description = "This is an indexed variable")
            Z, (description = "Scalar variable")
        end);

Note

Keywords must be contained in parentheses as in the example above.