JuMP.set_lower_bound

JuMP.set_start_value

JuMP.set_upper_bound

MPSGE.cost_function

MPSGE.extract_scalars

MPSGE.production_sectors

MPSGE.revenue_function

MPSGE.sectors

MPSGE.sectors

MPSGE.solve!

MPSGE.@auxiliaries

MPSGE.@auxiliary

MPSGE.@commodities

MPSGE.@commodity

MPSGE.@consumer

MPSGE.@consumers

MPSGE.@parameter

MPSGE.@parameters

MPSGE.@sector

MPSGE.@sectors

JuMP.set_lower_bound — Method

set_lower_bound(X::MPSGEScalarVariable, val::Real)

Set the lower bound of an MPSGE variable. This is an extension of the JuMP function set_lower_bound.

JuMP.set_start_value — Method

set_start_value(X::MPSGEScalarVariable, val::Real)

Set the staring value of an MPSGE variable. This is an extension of the JuMP function set_start_value.

JuMP.set_upper_bound — Method

set_upper_bound(X::MPSGEScalarVariable, val::Real)

Set the upper bound of an MPSGE variable. This is an extension of the JuMP function set_upper_bound.

MPSGE.cost_function — Method

cost_function(S::ScalarSector; virtual = false)
cost_function(S::ScalarSector, nest::Symbol; virtual = false)

Return a vector of cost functions for the given sector and nest. If nest is not provided return the cost function for input tree.

nest is the symbol representing the nest. This can also be the name of a commodity.

If virtual is true, return the virtual cost functions.

MPSGE.extract_scalars — Method

extract_scalars

Takes a variable and extracts it the sub-variables.

MPSGE.production_sectors — Method

production_sectors(m::MPSGEModel)

Return all sectors that have a corresponding production block. These are coming from a dictionary, so order is not guaranteed.

This is primarily used when generating constraints.

MPSGE.revenue_function — Method

revenue_function(S::ScalarSector; virtual = false)    
revenue_function(S::ScalarSector, nest::Symbol; virtual = false)

Return a vector of revenue functions for the given sector and nest. If nest is not provided return the revenue function for input tree.

nest is the symbol representing the nest. This can also be the name of a commodity.

If virtual is true, return the virtual revenue functions.

MPSGE.sectors — Method

sectors(C::Commodity)

Return only the sectors that have the input commodity in their production block.

This is an optimization in building the model as the structure is very sparse iterating over all sectors is expensive.

MPSGE.sectors — Method

sectors(m::MPSGEModel)

Return all sectors in a model

MPSGE.solve! — Method

solve!(m::abstract_mpsge_model; keywords)
Function to solve the model. Triggers the build if the model hasn't been built yet.

Example

julia> solve!(m, cumulative_iteration_limit=0)

MPSGE.@auxiliaries — Macro

@auxiliaries(model, args...)

Adds multiple @auxiliary to a model.

The model must be the first argument, and multiple auxiliaries can be added in a begin ... end block, one auxiliary per line.

Example

julia> M = MPSGEModel()

julia> I = [:a,:b]

julia> @auxiliaries(M, begin
            X[i=I]
            Y[i=I,j=I], (description = "This is an indexed variable")
            Z, (description = "Scalar variable")
        end)

Note

Keywords must be contained in parentheses as in the example above.

MPSGE.@auxiliary — Macro

@auxiliary(model, X, kwargs...)
@auxiliary(model, X[I], kwargs...)
@auxiliary(model, X[i=I], kwargs...)

Create a auxiliary in the model with name X, index I, and index name i with arguments kwargs.

Arguments

model: An isntance of MPSGEModel.

The auxiliary X can be instantiated in a several ways.

X creates a scalar auxiliary

If I is a defined array, then X[I] creates an indexed auxiliary with

indices given by I.

If I is a defined array, then X[i=I]creates an indexed auxiliary with

indices given by I and variable index named i.

If you want to create an index auxiliary it is highly recommended to use the syntax from (3) with better name for i, For example X[goods=I], this will get the index name to goods which will be displayed when printing the model. For example creating a variable using

@auxiliary(M, X[goods=I], description = "Auxiliary with indexed variables")

will allow the auxiliary to be printed as

X[goods] -- Auxiliary with indexed variables

Additionally, multi-indexed auxiliaries can be created by using the syntax X[regions = R, goods = G].

Optional Arguments

description: Set a description on a variable.

start: Set the starting value of the variable Default 0.0

lower_bound: Set the lower bound of the variable. Default -Inf.

upper_bound: Set the upper bound of the variable. Default Inf.

Examples

using MPSGE

R = Symbol.(:r, 1:5)
G = Symbol.(:g, 1:5)
M = MPSGEModel()

@auxiliary(M, X[region=R, goods=G], description="Auxiliary with indexed variables")

Note

By default auxiliary variables start at 0 with no lower or upper bounds. These can be set after variable creating using set_start_value, set_lower_bound, and set_upper_bound.

MPSGE.@commodities — Macro

@commodities(model, args...)

Adds multiple @commodity to a model.

The model must be the first argument, and multiple commodities can be added in a begin ... end block, one commodity per line.

Example

julia> M = MPSGEModel()

julia> I = [:a,:b]

julia> @commodities(M, begin
            X[i=I]
            Y[i=I,j=I], (description = "This is an indexed variable")
            Z, (description = "Scalar variable")
        end)

Note

Keywords must be contained in parentheses as in the example above.

MPSGE.@commodity — Macro

@commodity(model, C, kwargs...)
@commodity(model, C[I], kwargs...)
@commodity(model, C[i=I], kwargs...)

Create a commodity in the model with name C, index I, and index name i with arguments kwargs.

Arguments

model: An isntance of MPSGEModel.

The commodity C can be instantiated in a several ways.

C creates a scalar commodity

If I is a defined array, then C[I] creates an indexed commodity with

indices given by I.

If I is a defined array, then C[i=I]creates an indexed commodity with

indices given by I and variable index named i.

If you want to create an index commodity it is highly recommended to use the syntax from (3) with better name for i, For example C[goods=I], this will get the index name to goods which will be displayed when printing the model. For example creating a variable using

@commodity(M, C[goods=I], description = "Commodity with indexed variables")

will allow the commodity to be printed as

C[goods] -- Commodity with indexed variables

Additionally, multi-indexed commoditys can be created by using the syntax C[regions = R, goods = G].

Optional Arguments

description: Set a description on a variable.

start: Set the starting value of the variable Default 1.0

lower_bound: Set the lower bound of the variable. Default 0.

upper_bound: Set the upper bound of the variable. Default Inf.

Examples

using MPSGE

R = Symbol.(:r, 1:5)
G = Symbol.(:g, 1:5)
M = MPSGEModel()

@commodity(M, C[region=R, goods=G], description="Commodity with indexed variables")

MPSGE.@consumer — Macro

@consumer(model, H, kwargs...)
@consumer(model, H[I], kwargs...)
@consumer(model, H[i=I], kwargs...)

Create a consumer in the model with name H, index I, and index name i with arguments kwargs.

Arguments

model: An isntance of MPSGEModel.

The consumer H can be instantiated in a several ways.

H creates a scalar consumer

If I is a defined array, then H[I] creates an indexed consumer with

indices given by I.

If I is a defined array, then H[i=I]creates an indexed consumer with

indices given by I and variable index named i.

If you want to create an index consumer it is highly recommended to use the syntax from (3) with better name for i, For example H[goods=I], this will get the index name to goods which will be displayed when printing the model. For example creating a variable using

@consumer(M, H[goods=I], description = "Consumer with indexed variables")

will allow the consumer to be printed as

H[goods] -- Consumer with indexed variables

Additionally, multi-indexed consumers can be created by using the syntax H[regions = R, goods = G].

Optional Arguments

description: Set a description on a variable.

start: Set the starting value of the variable Default 1.0

lower_bound: Set the lower bound of the variable. Default 0.

upper_bound: Set the upper bound of the variable. Default Inf.

Examples

using MPSGE

R = Symbol.(:r, 1:5)
G = Symbol.(:g, 1:5)
M = MPSGEModel()

@consumer(M, H[region=R, goods=G], description="Consumer with indexed variables")

MPSGE.@consumers — Macro

@consumers(model, args...)

Adds multiple @consumer to a model.

The model must be the first argument, and multiple consumers can be added in a begin ... end block, one consumer per line.

Example

julia> M = MPSGEModel()

julia> I = [:a,:b]

julia> @consumers(M, begin
            X[i=I]
            Y[i=I,j=I], (description = "This is an indexed variable")
            Z, (description = "Scalar variable")
        end)

Note

Keywords must be contained in parentheses as in the example above.

MPSGE.@parameter — Macro

@parameter(model, expr, value, kwargs...)

Add a parameter to the model described by expr with initial value value and keyword arguments kwargs....

Required Arguments

model is the MPSGE model

expr an expression that describes the name and index of the parameter. See the examples below

value The initial value of the parameter.

Keyword Arguments

description A string describing the parameter.

Examples

Non-Indexed Parameters

using MPSGE

M = MPSGEModel()

initial_value = 10

@parameter(M, X, 1)
@parameter(M, Y, initial_value, description="A parameter")

Indexed Parameters

using MPSGE

M = MPSGEModel()

R = 1:5
S = 1:3

one_dimension = Dict(r => 2*r for r in R)
two_dimension = Dict((r,s) => r+s for r in R, s in S)

@parameter(M, X[r=R], 1, description = "example") # Index `R` and value 1 with a description
@parameter(M, Y[R, S], 1) # Indices `R` and `S` and value 1.
@parameter(M, Z[r=R], one_dimension[r]) # Index `R` and values `one_dimension[r]`.
@parameter(M, W[r=R, s=S], two_dimension[r, s]) # Indices `R` and `S` and values `two_dimension[r, s]`.

MPSGE.@parameters — Macro

@parameters(model, args...)

Adds multiple parameters to model at once, in the same fashion as the @parameter macro.

The model must be the first argument, and multiple parameters can be added on multiple lines wrapped in a begin ... end block.

The macro returns a tuple containing the parameters that were defined.

Example

using MPSGE
model = MPSGEModel();

@parameters(model, begin
           x, 1
           y[i = 1:2], i, (description = "y parameter")
       end)

Note

Keyword arguments must be contained within parentheses (refer to the example above).

MPSGE.@sector — Macro

@sector(model, S, kwargs...)
@sector(model, S[I], kwargs...)
@sector(model, S[i=I], kwargs...)

Create a sector in the model with name S, index I, and index name i with arguments kwargs.

Arguments

model: An isntance of MPSGEModel.

The sector S can be instantiated in a several ways.

S creates a scalar sector

If I is a defined array, then S[I] creates an indexed sector with

indices given by I.

If I is a defined array, then S[i=I]creates an indexed sector with

indices given by I and variable index named i.

If you want to create an index sector it is highly recommended to use the syntax from (3) with better name for i, For example S[goods=I], this will set the index name to goods which will be displayed when printing the model. For example creating a variable using

@sector(M, S[goods=I], description = "Sector with indexed variables")

will allow the sector to be printed as

S[goods] -- Sector with indexed variables

Additionally, multi-indexed sectors can be created by using the syntax S[regions = R, goods = G].

Optional Arguments

description: Set a description on a variable.

start: Set the starting value of the variable Default 1.0

lower_bound: Set the lower bound of the variable. Default 0.

upper_bound: Set the upper bound of the variable. Default Inf.

Examples

using MPSGE

R = Symbol.(:r, 1:5)
G = Symbol.(:g, 1:5)
M = MPSGEModel()

@sector(M, S[region=R, goods=G], description="Sector with indexed variables", start = 1.5)

MPSGE.@sectors — Macro

@sectors(model, args...)

Adds multiple @sector to a model.

The model must be the first argument, and multiple sectors can be added in a begin ... end block, one sector per line.

Example

julia> M = MPSGEModel()

julia> I = [:a,:b]

julia> @sectors(M, begin
            X[i=I]
            Y[i=I,j=I], (description = "This is an indexed variable")
            Z, (description = "Scalar variable")
        end)

Note

Keywords must be contained in parentheses as in the example above.