Reference Guide

Missing docstring.

Missing docstring for Model. Check Documenter's build log for details.

MPSGE.add! — Function

add!(m,bar)
Function that adds an element to the model with a name assignment
m::Model is always the first Argument

# Options
Parameter::ScalarParameter, ::IndexedParameter
Commodity::ScalarCommodity, ::IndexedCommodity
Sector::ScalarSector,       ::IndexedSector
Consumer::ScalarConsumer,   ::IndexedConsumer
Aux::ScalarAux,             ::IndexedAux

Example

julia> S = add!(m, Sector())

Production::Production
Demand::DemandFunction
AuxConstraint::AuxConstraint

Example

julia> add!(m, Production()) 

JuMP.set_value — Function

set_value(P, value::Float64)
Function that allows users to set a specific value for a variable, updating the benchmark field.

Options

Parameter::ScalarParameter, ::IndexedParameter
Commodity::ScalarCommodity, ::IndexedCommodity
Sector::ScalarSector,       ::IndexedSector
Consumer::ScalarConsumer,   ::IndexedConsumer
Aux::ScalarAux,             ::IndexedAux

Example

julia> set_value(var, 1.3)

MPSGE.set_fixed! — Function

set_fixed!(P, true::Boolean)
Function that allows users to fix a value for a variable, the benchmark, the value from set_value, or the previous value.

Options

Parameter::ScalarParameter, ::IndexedParameter
Commodity::ScalarCommodity, ::IndexedCommodity
Sector::ScalarSector,       ::IndexedSector
Consumer::ScalarConsumer,   ::IndexedConsumer
Aux::ScalarAux,             ::IndexedAux

Example

julia> set_value(var, false)

MPSGE.Parameter — Type

Parameter(:symbol; indices, value::Float64=1., string)
Struct that holds the name, indices if IndexedParameter, value, and optional description of a parameter within the model.

Options

Parameter::ScalarParameter, IndexedParameter

Example

julia> P = add!(Parameter(model, :P, value=1., description="Elasticity"))
julia> sectors = [:s1, :s2]
julia> P = add!(Parameter(model, :P, indices=(,sectors), value=1., description="Elasticity parameters for X Sector "))

Missing docstring.

Missing docstring for Sector. Check Documenter's build log for details.