Reference Guide

MPSGE.Model — Type

Model() The struct that stores all the elements of the model.

Example

julia> foo = Model()

MPSGE.add! — Function

add!(m,bar)
Function that adds an element to the model with a name assignment
m::Model is always the first Argument

# Options
Parameter::ScalarParameter, ::IndexedParameter
Commodity::ScalarCommodity, ::IndexedCommodity
Sector::ScalarSector,       ::IndexedSector
Consumer::ScalarConsumer,   ::IndexedConsumer
Aux::ScalarAux,             ::IndexedAux

Example

julia> S = add!(m, Sector())

Production::Production
Demand::DemandFunction
AuxConstraint::AuxConstraint

Example

julia> add!(m, Production()) 

MPSGE.Parameter — Type

Parameter(:symbol; indices, value::Float64=1., string)
Struct that holds the name, indices if IndexedParameter, value, and optional description of a parameter within the model.

Options

Parameter::ScalarParameter, IndexedParameter

Example

julia> P = add!(Parameter(model, :P, value=1., description="Elasticity"))
julia> sectors = [:s1, :s2]
julia> P = add!(Parameter(model, :P, indices=(,sectors), value=1., description="Elasticity parameters for X Sector "))

MPSGE.Sector — Type

Sector(:symbol; indices, value::Float64=1., string)
Struct that holds the name, (indices if IndexedSector), value, and optional description of a sector within the model.

Options

Sector::ScalarSector, IndexedSector

Example

julia> S = add!(Sector(model, :S, value=1., description="Sector S"))
julia> sectors = [:s1, :s2]
julia> P = add!(Sector(model, :S, indices=(,sectors), value=1., description="S[:s1] and S[:s2] Sectors"))

MPSGE.Input — Type

Input(inputname::Symbol, value::Float64; taxes=taxes::Vector{Tax}, price=price::Union{Float64,Expr}=1.) The struct that stores all the elements of an Input.

Options

    Taxes and price are optional, keyword must be used.

Example

julia> Input(:PL, 50, taxes=[Tax(1., RA)], price=1.2)

MPSGE.Output — Type

Output(outputname::Symbol, value::Float64; taxes=taxes::Vector{Tax}, price=price::Union{Float64,Expr}=1.) The struct that stores all the elements of an Input.

Options

    Taxes and price are optional, keyword must be used.

Example

julia> Output(:PU, 50, taxes=[Tax(0.1, CONS)], price=.9)

JuMP.set_value — Function

set_value(P, value::Float64)
Function that allows users to set a specific value for a variable, updating the benchmark field.

Options

Parameter::ScalarParameter, ::IndexedParameter
Commodity::ScalarCommodity, ::IndexedCommodity
Sector::ScalarSector,       ::IndexedSector
Consumer::ScalarConsumer,   ::IndexedConsumer
Aux::ScalarAux,             ::IndexedAux

Example

julia> set_value(var, 1.3)

MPSGE.set_fixed! — Function

set_fixed!(P, true::Boolean)
Function that allows users to fix a value for a variable, the benchmark, the value from set_value, or the previous value.

Options

Parameter::ScalarParameter, ::IndexedParameter
Commodity::ScalarCommodity, ::IndexedCommodity
Sector::ScalarSector,       ::IndexedSector
Consumer::ScalarConsumer,   ::IndexedConsumer
Aux::ScalarAux,             ::IndexedAux

Example

julia> set_fixed!(var, false)